fix: Revert "refactor: move Slack client functionality to service layer"
This reverts commit 7abd151fe19fb5649d8cae4c85f4e8d9e59f758e.
