+++ packages/backend/src/clients/Slack/SlackMessageBlocks.ts
@@ -11,6 +11,7 @@ import {
     SectionBlock,
     SectionBlockAccessory,
 } from '@slack/bolt';
+import { Unfurl } from '../../services/UnfurlService/UnfurlService';
 import { AttachmentUrl } from '../EmailClient/EmailClient';
 
 type GetChartAndDashboardBlocksArgs = {
@@ -430,19 +431,6 @@ const getExploreBlocks = (
               }
             : undefined,
     ]);
-
-export type Unfurl = {
-    title: string;
-    description?: string;
-    chartType?: string;
-    imageUrl: string | undefined;
-    pageType: LightdashPage;
-    minimalUrl: string;
-    organizationUuid: string;
-    resourceUuid: string | undefined;
-    chartTileUuids?: (string | null)[];
-    sqlChartTileUuids?: (string | null)[];
-};
 
 export const getUnfurlBlocks = (
     originalUrl: string,
