+++ packages/backend/src/config/parseConfig.ts
@@ -245,13 +245,7 @@ export const getPemFileContent = (certValue: string | undefined) =>
         decodeUnlessStartsWith: '-----BEGIN ', // -----BEGIN CERTIFICATE | -----BEGIN PRIVATE KEY
     });
 
-export type LoggingLevel =
-    | 'error'
-    | 'warn'
-    | 'info'
-    | 'http'
-    | 'debug'
-    | 'audit';
+type LoggingLevel = 'error' | 'warn' | 'info' | 'http' | 'debug' | 'audit';
 const assertIsLoggingLevel = (x: string): x is LoggingLevel =>
     ['error', 'warn', 'info', 'http', 'debug', 'audit'].includes(x);
 const parseLoggingLevel = (raw: string): LoggingLevel => {
