+++ packages/backend/src/ee/index.ts
@@ -37,6 +37,7 @@ type EnterpriseAppArguments = Pick<
     | 'serviceProviders'
     | 'modelProviders'
     | 'customExpressMiddlewares'
+    | 'slackClientFactory'
 >;
 
 export async function getEnterpriseAppArguments(): Promise<EnterpriseAppArguments> {
@@ -98,12 +99,6 @@ export async function getEnterpriseAppArguments(): Promise<EnterpriseAppArgument
                     asyncQueryService: repository.getAsyncQueryService(),
                     userAttributesModel: models.getUserAttributesModel(),
                     searchModel: models.getSearchModel(),
-                    spaceModel: models.getSpaceModel(),
-                    slackAuthenticationModel:
-                        models.getSlackAuthenticationModel() as CommercialSlackAuthenticationModel,
-                    schedulerClient:
-                        clients.getSchedulerClient() as CommercialSchedulerClient,
-                    openIdIdentityModel: models.getOpenIdIdentityModel(),
                     spaceService: repository.getSpaceService(),
                 }),
             rolesService: ({ repository, context, models }) =>
@@ -320,6 +315,18 @@ export async function getEnterpriseAppArguments(): Promise<EnterpriseAppArgument
                 asyncQueryService:
                     context.serviceRepository.getAsyncQueryService(),
             }),
+        slackClientFactory: (context) =>
+            new CommercialSlackClient({
+                lightdashConfig: context.lightdashConfig,
+                analytics: context.analytics,
+                slackAuthenticationModel:
+                    context.models.getSlackAuthenticationModel() as CommercialSlackAuthenticationModel,
+                schedulerClient:
+                    context.clients.getSchedulerClient() as CommercialSchedulerClient,
+                aiAgentModel: context.models.getAiAgentModel(),
+                openIdIdentityModel: context.models.getOpenIdIdentityModel(),
+                userModel: context.models.getUserModel(),
+            }),
         clientProviders: {
             schedulerClient: ({ context, models }) =>
                 new CommercialSchedulerClient({
@@ -327,12 +334,17 @@ export async function getEnterpriseAppArguments(): Promise<EnterpriseAppArgument
                     analytics: context.lightdashAnalytics,
                     schedulerModel: models.getSchedulerModel(),
                 }),
-            slackClient: ({ context, models }) =>
+            slackClient: ({ context, models, repository }) =>
                 new CommercialSlackClient({
                     analytics: context.lightdashAnalytics,
                     lightdashConfig: context.lightdashConfig,
                     slackAuthenticationModel:
                         models.getSlackAuthenticationModel() as CommercialSlackAuthenticationModel,
+                    schedulerClient:
+                        repository.getSchedulerClient() as CommercialSchedulerClient,
+                    aiAgentModel: models.getAiAgentModel(),
+                    openIdIdentityModel: models.getOpenIdIdentityModel(),
+                    userModel: models.getUserModel(),
                 }),
         },
     };
