+++ packages/backend/src/services/ServiceRepository.ts
@@ -675,10 +675,6 @@ export class ServiceRepository
                     s3Client: this.clients.getS3Client(),
                     projectModel: this.models.getProjectModel(),
                     downloadFileModel: this.models.getDownloadFileModel(),
-                    slackClient: this.clients.getSlackClient(),
-                    analytics: this.context.lightdashAnalytics,
-                    slackAuthenticationModel:
-                        this.models.getSlackAuthenticationModel(),
                 }),
         );
     }
