+++ packages/backend/src/services/UnfurlService/UnfurlService.ts
@@ -9,46 +9,27 @@ import {
     getErrorMessage,
     isDashboardChartTileType,
     isDashboardSqlChartTile,
-    LightdashMode,
     LightdashPage,
     LightdashRequestMethodHeader,
     QueryHistoryStatus,
     RequestMethod,
     ScreenshotError,
     SessionUser,
-    SlackInstallationNotFoundError,
     snakeCaseName,
 } from '@lightdash/common';
 import * as Sentry from '@sentry/node';
-import {
-    AllMiddlewareArgs,
-    App,
-    LinkSharedEvent,
-    SlackEventMiddlewareArgs,
-} from '@slack/bolt';
-import { StringIndexed } from '@slack/bolt/dist/types/helpers';
-import { WebClient } from '@slack/web-api';
 import * as fsPromise from 'fs/promises';
 import { nanoid as useNanoid } from 'nanoid';
 import fetch from 'node-fetch';
 import { PDFDocument } from 'pdf-lib';
 import playwright from 'playwright';
-import { LightdashAnalytics } from '../../analytics/LightdashAnalytics';
 import { S3Client } from '../../clients/Aws/S3Client';
-import { SlackClient } from '../../clients/Slack/SlackClient';
-import {
-    getUnfurlBlocks,
-    Unfurl,
-} from '../../clients/Slack/SlackMessageBlocks';
 import { LightdashConfig } from '../../config/parseConfig';
-import { slackErrorHandler } from '../../errors';
-import Logger from '../../logging/logger';
 import { DashboardModel } from '../../models/DashboardModel/DashboardModel';
 import { DownloadFileModel } from '../../models/DownloadFileModel';
 import { ProjectModel } from '../../models/ProjectModel/ProjectModel';
 import { SavedChartModel } from '../../models/SavedChartModel';
 import { ShareModel } from '../../models/ShareModel';
-import { SlackAuthenticationModel } from '../../models/SlackAuthenticationModel';
 import { SpaceModel } from '../../models/SpaceModel';
 import { getAuthenticationToken } from '../../routers/headlessBrowser';
 import { BaseService } from '../BaseService';
@@ -79,6 +60,19 @@ export enum ScreenshotContext {
 
 const SCREENSHOT_RETRIES = 3;
 
+export type Unfurl = {
+    title: string;
+    description?: string;
+    chartType?: string;
+    imageUrl: string | undefined;
+    pageType: LightdashPage;
+    minimalUrl: string;
+    organizationUuid: string;
+    resourceUuid: string | undefined;
+    chartTileUuids?: (string | null)[];
+    sqlChartTileUuids?: (string | null)[];
+};
+
 export type ParsedUrl = {
     isValid: boolean;
     lightdashPage?: LightdashPage;
@@ -90,33 +84,6 @@ export type ParsedUrl = {
     exploreModel?: string;
 };
 
-const notifySlackError = async (
-    error: unknown,
-    url: string,
-    client: WebClient,
-    event: LinkSharedEvent,
-    { appProfilePhotoUrl }: { appProfilePhotoUrl?: string },
-): Promise<void> => {
-    /** Expected slack errors:
-     * - cannot_parse_attachment: Means the image on the blocks is not accessible from slack, is the URL public ?
-     */
-    Logger.error(`Unable to unfurl slack URL ${url}: ${error} `);
-
-    // Send message in thread
-    await client.chat
-        .postMessage({
-            thread_ts: event.message_ts,
-            channel: event.channel,
-            ...(appProfilePhotoUrl ? { icon_url: appProfilePhotoUrl } : {}),
-            text: `:fire: Unable to unfurl ${url}: ${error}`,
-        })
-        .catch((er: unknown) =>
-            Logger.error(
-                `Unable send slack error message: ${getErrorMessage(er)}`,
-            ),
-        );
-};
-
 type UnfurlServiceArguments = {
     lightdashConfig: LightdashConfig;
     dashboardModel: DashboardModel;
@@ -124,11 +91,8 @@ type UnfurlServiceArguments = {
     spaceModel: SpaceModel;
     shareModel: ShareModel;
     s3Client: S3Client;
-    slackClient: SlackClient;
     projectModel: ProjectModel;
     downloadFileModel: DownloadFileModel;
-    analytics: LightdashAnalytics;
-    slackAuthenticationModel: SlackAuthenticationModel;
 };
 
 export class UnfurlService extends BaseService {
@@ -144,15 +108,9 @@ export class UnfurlService extends BaseService {
 
     s3Client: S3Client;
 
-    slackClient: SlackClient;
-
     projectModel: ProjectModel;
 
     downloadFileModel: DownloadFileModel;
-
-    analytics: LightdashAnalytics;
-
-    slackAuthenticationModel: SlackAuthenticationModel;
 
     constructor({
         lightdashConfig,
@@ -163,9 +121,6 @@ export class UnfurlService extends BaseService {
         s3Client,
         projectModel,
         downloadFileModel,
-        slackClient,
-        analytics,
-        slackAuthenticationModel,
     }: UnfurlServiceArguments) {
         super();
         this.lightdashConfig = lightdashConfig;
@@ -174,46 +129,8 @@ export class UnfurlService extends BaseService {
         this.spaceModel = spaceModel;
         this.shareModel = shareModel;
         this.s3Client = s3Client;
-        this.slackClient = slackClient;
         this.projectModel = projectModel;
         this.downloadFileModel = downloadFileModel;
-        this.analytics = analytics;
-        this.slackAuthenticationModel = slackAuthenticationModel;
-
-        this.initSlackListeners();
-    }
-
-    private initSlackListeners() {
-        if (!this.slackClient.isEnabled) return;
-
-        if (this.slackClient.isReady()) {
-            const slackApp = this.slackClient.getApp();
-            if (slackApp) {
-                this.registerSlackEventHandlers(slackApp);
-            }
-        } else {
-            this.slackClient.once('slackAppReady', (slackApp: App) => {
-                this.registerSlackEventHandlers(slackApp);
-            });
-        }
-    }
-
-    private registerSlackEventHandlers(slackApp: App) {
-        try {
-            slackApp.event(
-                'link_shared',
-                (
-                    m: SlackEventMiddlewareArgs<'link_shared'> &
-                        AllMiddlewareArgs<StringIndexed>,
-                ) => this.unfurlSlackUrls(m),
-            );
-            Logger.info('UnfurlService: Slack link_shared listener registered');
-        } catch (error) {
-            Logger.error(
-                'UnfurlService: Failed to register Slack listeners:',
-                error,
-            );
-        }
     }
 
     private async waitForAllPaginatedResultsResponse(
@@ -1259,132 +1176,5 @@ export class UnfurlService extends BaseService {
             );
         }
         return header;
-    }
-
-    private async sendUnfurl(
-        event: LinkSharedEvent,
-        originalUrl: string,
-        unfurl: Unfurl,
-        client: WebClient,
-    ) {
-        const unfurlBlocks = getUnfurlBlocks(originalUrl, unfurl);
-        await client.chat
-            .unfurl({
-                ts: event.message_ts,
-                channel: event.channel,
-                unfurls: unfurlBlocks,
-            })
-            .catch((e: unknown) => {
-                this.analytics.track({
-                    event: 'share_slack.unfurl_error',
-                    userId: event.user,
-                    properties: {
-                        error: `${getErrorMessage(e)}`,
-                    },
-                });
-                Logger.error(
-                    `Unable to unfurl on slack ${JSON.stringify(
-                        unfurlBlocks,
-                    )}: ${JSON.stringify(e)}`,
-                );
-            });
-    }
-
-    private async unfurlSlackUrls(
-        message: SlackEventMiddlewareArgs<'link_shared'> &
-            AllMiddlewareArgs<StringIndexed>,
-    ) {
-        const { event, client, context } = message;
-        let appProfilePhotoUrl: string | undefined;
-
-        if (event.channel === 'COMPOSER') return; // Do not unfurl urls when typing, only when message is sent
-
-        Logger.debug(`Got link_shared slack event ${event.message_ts}`);
-
-        event.links.map(async (l) => {
-            const eventUserId = context.botUserId;
-
-            try {
-                const { teamId } = context;
-                const details = await this.unfurlDetails(l.url);
-
-                if (details) {
-                    this.analytics.track({
-                        event: 'share_slack.unfurl',
-                        userId: eventUserId,
-                        properties: {
-                            organizationId: details?.organizationUuid,
-                        },
-                    });
-
-                    Logger.debug(
-                        `Unfurling ${details.pageType} with URL ${details.minimalUrl}`,
-                    );
-
-                    await this.sendUnfurl(event, l.url, details, client);
-
-                    const imageId = `slack-image-${useNanoid()}`;
-                    const authUserUuid =
-                        await this.slackAuthenticationModel.getUserUuid(
-                            teamId ?? '',
-                        );
-
-                    const installation =
-                        await this.slackAuthenticationModel.getInstallationFromOrganizationUuid(
-                            details?.organizationUuid,
-                        );
-
-                    if (!installation) {
-                        throw new SlackInstallationNotFoundError();
-                    }
-
-                    appProfilePhotoUrl = installation.appProfilePhotoUrl;
-
-                    const { imageUrl } = await this.unfurlImage({
-                        url: details.minimalUrl,
-                        lightdashPage: details.pageType,
-                        imageId,
-                        authUserUuid,
-                        context: ScreenshotContext.SLACK,
-                    });
-
-                    if (imageUrl) {
-                        await this.sendUnfurl(
-                            event,
-                            l.url,
-                            { ...details, imageUrl },
-                            client,
-                        );
-
-                        this.analytics.track({
-                            event: 'share_slack.unfurl_completed',
-                            userId: eventUserId,
-                            properties: {
-                                pageType: details.pageType,
-                                organizationId: details?.organizationUuid,
-                            },
-                        });
-                    }
-                }
-            } catch (e) {
-                if (this.lightdashConfig.mode === LightdashMode.PR) {
-                    void notifySlackError(e, l.url, client, event, {
-                        appProfilePhotoUrl,
-                    });
-                }
-                if (!(e instanceof ScreenshotError)) {
-                    slackErrorHandler(e, 'Unable to unfurl slack URL');
-                }
-
-                this.analytics.track({
-                    event: 'share_slack.unfurl_error',
-                    userId: eventUserId,
-
-                    properties: {
-                        error: `${e}`,
-                    },
-                });
-            }
-        });
     }
 }
